Renamed 'umpost' to 'um2nc' and moved to top-level 'src' folder.
