+++ .gitattributes
@@ -1,1 +1,1 @@
-umpost/_version.py export-subst
+um2nc/_version.py export-subst
