+++ integration/binary_diff.sh
@@ -68,7 +68,7 @@ mask_path=$UM2NC_TEST_DATA/mask.nc
 # execute nomask variant, pressure masking is turned OFF & all cubes are kept
 # TODO: capture error condition if conversion does not complete
 rm -f "$nomask_path"  # remove previous data
-python3 "$UM2NC_PROJ"/umpost/um2netcdf.py --verbose \
+python3 "$UM2NC_PROJ"/um2nc/um2netcdf.py --verbose \
                                         --nohist \
                                         --nomask \
                                         "$subset_path" \
@@ -80,7 +80,7 @@ echo
 # execute pressure masking variant: cubes which cannot be pressure masked are dropped
 # TODO: capture error condition if conversion does not complete
 rm -f "$mask_path"  # remove previous data
-python3 "$UM2NC_PROJ"/umpost/um2netcdf.py --verbose \
+python3 "$UM2NC_PROJ"/um2nc/um2netcdf.py --verbose \
                                         --nohist \
                                         "$subset_path" \
                                         "$mask_path"
