+++ src/um2nc/__init__.py
@@ -1,3 +1,3 @@
-from umpost import _version
+from um2nc import _version
 
 __version__ = _version.get_versions()["version"]
